UserDetail front 쪽 dto -> JSON으로 정보 전달
